Added class BreedRepository Added routers to handles '/cats' and '/breeds' path requests seperately
